requirements.txt
Django==1.9.6
djangorestframework==3.3.3
markdown==2.6.6
django-filter==0.13.0
